Remove unused files; split new code into multiple files
