+++ tuner/src/main/scala/org/calinburloiu/music/microtonalist/tuner/TuningChanger.scala
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021 Calin-Andrei Burloiu
+ * Copyright 2025 Calin-Andrei Burloiu
  *
  *    Licensed under the Apache License, Version 2.0 (the "License");
  *    you may not use this file except in compliance with the License.
@@ -16,6 +16,12 @@
 
 package org.calinburloiu.music.microtonalist.tuner
 
-import org.calinburloiu.music.scmidi.MidiProcessor
+import org.calinburloiu.music.microtonalist.common.Plugin
 
-trait TuningSwitchProcessor extends MidiProcessor {}
+import javax.sound.midi.MidiMessage
+
+abstract class TuningChanger extends Plugin {
+  override val familyName: String = "tuningChanger"
+
+  def decide(message: MidiMessage): TuningChange
+}
